+++ examples/globus_compute/README.md
@@ -132,7 +132,7 @@ To set up a federated learning server using Globus Compute for real-world FL exp
 - `get_loss`: function to get the training loss function
 - `val_metric`: function to validate the trained model
 
-6. The server also needs to collect the following information and files from the clients, and put them in the [client configuration file](./configs_client/mnist.yaml).
+6. The server also needs to collect the following information and files from the clients, and put them in the [client configuration file](./configs_client/mnist_iid.yaml).
 - `endpoint_id`: Globus Compute Endpoint ID
 - `device`: computing device for the client, `cpu` or `cuda`
 - `get_data`: data loader function to load the client local dataset. [**Note**: This function will be run on the client computing machine, on the directory where the client starts the Globus Compute endpoint, so the client has to make sure that the data path in the data loader file is correct.]
@@ -161,6 +161,6 @@ Explanations for the provided server configuration files in `configs_server`:
 - [`mnist_fedcompass_step_optim.yaml`](configs_server/mnist_fedcompass_step_optim.yaml): Server configuration for [FedCompass](https://arxiv.org/abs/2309.14675) algorithm, and each client updates for different number of local steps according to their computing power for each local training round.
 
 Explanations for the provided client configuration files in `configs_client`:
-- [`mnist.yaml`](configs_client/mnist.yaml): The client local datasets are equally and randomly partitioned MNIST dataset, which is identically and independently distributed (IID).
+- [`mnist_iid.yaml`](configs_client/mnist_iid.yaml): The client local datasets are equally and randomly partitioned MNIST dataset, which is identically and independently distributed (IID).
 - [`mnist_class_noiid.yaml`](configs_client/mnist_class_noiid.yaml): The client local datasets are non-IID partitioned MNIST dataset using the *Class Partition* strategy, check **Appendix D.1.1** of our [FedCompass](https://arxiv.org/pdf/2309.14675.pdf) paper for details.
 - [`mnist_dual_dirichlet_noiid.yaml`](configs_client/mnist_dual_dirichlet_noiid.yaml): The client local datasets are non-IID partitioned MNIST dataset using the *Dual Dirichlet Partition* strategy, check **Appendix D.1.2** of our [FedCompass](https://arxiv.org/pdf/2309.14675.pdf) paper for details.
