+++ examples/globus_compute/configs_client/mnist_dual_dirichlet_noiid.yaml
@@ -3,7 +3,7 @@ clients:
     endpoint_id: "6c87988d-d067-43c9-a73f-063b51e1b33a"
     device:  cpu
     output_dir: "./output"
-    data_dir:   "./output"
+    data_dir:   "./datasets"
     get_data:
       script_file: "globus_compute/datasets/mnist_dual_dirichlet_noiid.py"
       call:   "get_data"
@@ -12,7 +12,7 @@ clients:
     endpoint_id: "4d34697f-f2e0-41dc-87b4-9821a02c0c0a"
     device:  cpu
     output_dir: "./output"
-    data_dir:   "./output"
+    data_dir:   "./datasets"
     get_data:
       script_file: "globus_compute/datasets/mnist_dual_dirichlet_noiid.py"
       call:   "get_data"
@@ -21,7 +21,7 @@ clients:
     endpoint_id: "83e63530-c674-4de0-a209-d87ea9ed17fd"
     device:  cpu
     output_dir: "./output"
-    data_dir:   "./output"
+    data_dir:   "./datasets"
     get_data:
       script_file: "globus_compute/datasets/mnist_dual_dirichlet_noiid.py"
       call:   "get_data"
